- rewrite "if else" wall of text via typescript shenanigans - merge state and handler functions to the single class, those functions was using jippityState a lot so: methods > non-pure functions - moved new handler to separate file - added types here and there
